// app/admin/products/[productId]/interface.ts
export interface ResponseProduct {
  abilities:                Ability[];
  base_experience:          number;
  cries:                    Cries;
  forms:                    Species[];
  game_indices:             GameIndex[];
  height:                   number;
  held_items:               HeldItem[];
  id:                       number;
  is_default:               boolean;
  location_area_encounters: string;
  moves:                    Move[];
  name:                     string;
  order:                    number;
  past_abilities:           any[];
  past_types:               any[];
  species:                  Species;
  sprites:                  Sprites;
  stats:                    Stat[];
  types:                    Type[];
  weight:                   number;
 }
 
 export interface Ability {
  ability:   Species;
  is_hidden: boolean;
  slot:      number;
 }
 
 export interface Species {
  name: string;
  url:  string;
 }
 
 export interface Cries {
  latest: string;
  legacy: string;
 }
 
 export interface GameIndex {
  game_index: number;
  version:    Species;
 }
 
 export interface HeldItem {
  item:            Species;
  version_details: VersionDetail[];
 }
 
 export interface VersionDetail {
  rarity:  number;
  version: Species;
 }
 
 export interface Move {
  move:                  Species;
  version_group_details: VersionGroupDetail[];
 }
 
 export interface VersionGroupDetail {
  level_learned_at:  number;
  move_learn_method: Species;
  version_group:     Species;
 }
 
 export interface GenerationV {
  "black-white": Sprites;
 }
 
 export interface GenerationIv {
  "diamond-pearl":        Sprites;
  "heartgold-soulsilver": Sprites;
  platinum:               Sprites;
 }
 
 export interface Versions {
  "generation-i":    GenerationI;
  "generation-ii":   GenerationIi;
  "generation-iii":  GenerationIii;
  "generation-iv":   GenerationIv;
  "generation-v":    GenerationV;
  "generation-vi":   { [key: string]: Home };
  "generation-vii":  GenerationVii;
  "generation-viii": GenerationViii;
 }
 
 export interface Other {
  dream_world:        DreamWorld;
  home:               Home;
  "official-artwork": OfficialArtwork;
  showdown:           Sprites;
 }
 
 export interface Sprites {
  animated?:          Sprites;
  back_default:       string;
  back_female:        null;
  back_shiny:         string;
  back_shiny_female:  null;
  front_default:      string;
  front_female:       null;
  front_shiny:        string;
  front_shiny_female: null;
  other?:             Other;
  versions?:          Versions;
 }
// ...
 export interface Stat {
  base_stat: number;
  effort:    number;
  stat:      Species;
 }
 
 export interface Type {
  slot: number;
  type: Species;
 }